Changed int to BigDecimal in the program

+++ src/main/java/com/transaction/TransactionMicroService/Controllers/AccountRestController.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/account")
 public class AccountRestController {
@@ -24,13 +26,13 @@ public class AccountRestController {
     }
 
     @PostMapping("/credit/{id}")
-    public ResponseEntity<String> credit(@PathVariable Long id, @RequestParam int amount){
+    public ResponseEntity<String> credit(@PathVariable Long id, @RequestParam BigDecimal amount){
         transactionService.credit(id, amount);
         return ResponseEntity.ok("Credit transaction completed successfully");
     }
     //Returns different response depending if the transaction was successful or not, based on your balance
     @PostMapping("/debit/{id}")
-    public ResponseEntity<String> debit(@PathVariable Long id, @RequestParam int amount){
+    public ResponseEntity<String> debit(@PathVariable Long id, @RequestParam BigDecimal amount){
         boolean success = transactionService.debit(id, amount);
         if (success){
             return ResponseEntity.ok("Debit transaction completed successfully");
